requirements.txt
asgiref==3.6.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django_debug_toolbar==3.8.1
Pillow==9.4.0
sqlparse==0.4.3
